example/sanic_auth_test/requirements.txt
sanic
python-dotenv
ikon-auth>=0.1.5
